refactor(app): restructure project folders, modularize routes, and enhance maintainability
- Major folder restructuring for improved modularity and Clean Architecture adherence:
  - Split monolithic files in app/api, app/core, and app/infrastructure into submodules by feature/domain.
  - Moved and modularized API routes:
    - app/api/routes.py deleted, replaced with app/api/routes/ and submodules for auth, images, posters, and utils.
    - Added dedicated route files: auth/admin.py, auth/login.py, auth/register.py, images.py, posters.py, utils.py.
  - Refactored core entities and services:
    - app/core/entities.py and app/core/services.py deleted.
    - Introduced app/core/entities/ and app/core/services/ submodules for album, image, poster, token, user, and related services.
  - Infrastructure layer split:
    - app/infrastructure/postgresql_repositories.py and repositories.py deleted.
    - Added app/infrastructure/repositories/ with separate files for image, poster, token, and user repositories.
  - Added/updated config, exceptions, and utils modules for better separation of concerns.
- Updated and improved exception handling (app/exceptions/).
- Updated database and environment setup scripts; added album migration SQL and environment setup docs.
- Updated .gitignore to reflect new structure and ignore patterns.
- Refactored main.py:
  - Updated imports and app initialization to match new structure.
  - Enhanced OpenAPI documentation and middleware.
  - Improved static file and SPA serving logic.
  - Added/updated health check and WebSocket endpoints.
- Removed obsolete test files and env.example, added env.template.
- Updated Docker and database setup files for new structure.
- General code cleanup, improved maintainability, and better adherence to Clean Architecture principles.

BREAKING CHANGE: Project structure and import paths have changed. Update any custom scripts, deployment, or documentation referencing old paths.
