+++ .gitignore
@@ -23,7 +23,9 @@ ssh_config.backup
 docker/postgres/data/
 docker/pgadmin/data/
 backups/
+# Exclude SQL files except migration scripts
 *.sql
+!database/*.sql
 *.dump
 
 # Python
@@ -209,3 +211,10 @@ frontend/node_modules/
 # (Optional) Frontend environment files
 frontend/.env
 frontend/.env.local
+
+# Tests
+/tests/
+/tests/test_client.py
+/tests/test_expired_token.py
+/tests/test_refresh_token.py
+/tests/test_image.jpg
