+++ app/bootstrap.py
@@ -1,0 +1,3 @@
+from dotenv import load_dotenv
+
+load_dotenv()
