+++ app/config/__init__.py
@@ -1,0 +1,7 @@
+"""
+Configuration management
+"""
+
+from .settings import settings
+
+__all__ = ["settings"]
