+++ app/core/services/__init__.py
@@ -1,0 +1,18 @@
+"""
+Core Services Package
+Contains all business logic services organized by domain
+"""
+
+from .album_service import AlbumService
+from .auth_service import AuthService
+from .email_service import EmailService
+from .image_service import ImageService
+from .poster_service import PosterService
+
+__all__ = [
+    "AuthService",
+    "EmailService",
+    "ImageService",
+    "PosterService",
+    "AlbumService",
+]
