+++ app/infrastructure/repositories/__init__.py
@@ -1,0 +1,22 @@
+"""
+PostgreSQL Repository implementations
+Contains all repository classes organized by domain
+"""
+
+from ...utils.storage import LocalFileStorage
+from .image_repository import PostgreSQLImageRepository
+from .poster_repository import (PostgreSQLArchivedPosterRepository,
+                                PostgreSQLPosterRepository)
+from .token_repository import (PostgreSQLRefreshTokenRepository,
+                               PostgreSQLTokenRepository)
+from .user_repository import PostgreSQLUserRepository
+
+__all__ = [
+    "PostgreSQLUserRepository",
+    "PostgreSQLImageRepository",
+    "PostgreSQLTokenRepository",
+    "PostgreSQLRefreshTokenRepository",
+    "PostgreSQLPosterRepository",
+    "PostgreSQLArchivedPosterRepository",
+    "LocalFileStorage",
+]
