+++ database/database_setup.py
@@ -2,9 +2,10 @@
 """
 Minimal Database Setup Script
 - Creates tables
-- Creates default admin user (admin/admin123)
+- Creates default admin user from environment variables
 """
 import asyncio
+import os
 from datetime import datetime
 
 from dotenv import load_dotenv
@@ -12,7 +13,7 @@ from passlib.context import CryptContext
 
 from app.core.entities import User, UserStatus
 from app.infrastructure.database import AsyncSessionLocal, init_db
-from app.infrastructure.postgresql_repositories import PostgreSQLUserRepository
+from app.infrastructure.repositories import PostgreSQLUserRepository
 
 load_dotenv()
 
@@ -22,7 +23,7 @@ async def setup_database():
     await init_db()
     print("✅ Tables created successfully")
     await create_default_admin()
-    print("✅ Default admin user created (admin/admin123)")
+    print("✅ Default admin user created")
     print("🎉 Database setup completed!")
 
 
@@ -30,14 +31,21 @@ async def create_default_admin():
     pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
     async with AsyncSessionLocal() as session:
         user_repo = PostgreSQLUserRepository(session)
-        existing_user = await user_repo.get_by_username("admin")
+
+        # Get admin credentials from environment
+        admin_username = os.getenv("ADMIN_USERNAME", "admin")
+        admin_email = os.getenv("ADMIN_EMAIL", "admin@example.com")
+        admin_password = os.getenv("ADMIN_PASSWORD", "admin123")
+
+        existing_user = await user_repo.get_by_username(admin_username)
         if existing_user:
-            print("ℹ️ Admin user already exists")
+            print(f"ℹ️ Admin user '{admin_username}' already exists")
             return
+
         admin_user = User(
-            username="admin",
-            email="admin@example.com",
-            hashed_password=pwd_context.hash("admin123"),
+            username=admin_username,
+            email=admin_email,
+            hashed_password=pwd_context.hash(admin_password),
             is_active=True,
             is_admin=True,
             status=UserStatus.APPROVED,
@@ -45,7 +53,7 @@ async def create_default_admin():
             approved_by="system",
         )
         await user_repo.create(admin_user)
-        print("👤 Created admin user: admin/admin123")
+        print(f"👤 Created admin user: {admin_username}/{admin_password}")
 
 
 def main():
