--- env.example
@@ -1,32 +1,0 @@
-# Database Configuration for Docker PostgreSQL
-DATABASE_URL=postgresql+asyncpg://postgres:password@localhost:5432/image_upload_db
-DB_ECHO=false
-
-# Security Configuration
-SECRET_KEY=your-super-secret-key-change-this-in-production
-ALGORITHM=HS256
-ACCESS_TOKEN_EXPIRE_MINUTES=30
-REFRESH_TOKEN_EXPIRE_DAYS=7
-
-# Email Configuration for Admin Notifications
-MAIL_USERNAME=your-email@gmail.com
-MAIL_PASSWORD=your-app-password
-MAIL_FROM=your-email@gmail.com
-MAIL_PORT=587
-MAIL_SERVER=smtp.gmail.com
-ADMIN_EMAIL=admin@example.com
-
-# Server Configuration
-HOST=0.0.0.0
-PORT=8000
-
-# File Upload Settings
-MAX_FILE_SIZE=10485760  # 10MB in bytes
-ALLOWED_IMAGE_TYPES=image/jpeg,image/png,image/gif,image/webp
-
-# Docker Configuration
-POSTGRES_HOST=localhost
-POSTGRES_PORT=5432
-POSTGRES_DB=image_upload_db
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=password
