+++ env.template
@@ -1,0 +1,49 @@
+# Database Configuration
+DATABASE_URL=postgresql+asyncpg://user:password@localhost:5432/linklink
+DB_ECHO=false
+
+# Security Configuration
+SECRET_KEY=your-secret-key-change-this-in-production
+ACCESS_TOKEN_EXPIRE_MINUTES=30
+REFRESH_TOKEN_EXPIRE_DAYS=7
+
+# Email Configuration for Admin Notifications
+MAIL_USERNAME=your-email@gmail.com
+MAIL_PASSWORD=your-app-password
+MAIL_FROM=your-email@gmail.com
+MAIL_PORT=587
+MAIL_SERVER=smtp.gmail.com
+MAIL_STARTTLS=true
+MAIL_SSL_TLS=false
+USE_CREDENTIALS=true
+ADMIN_EMAIL=admin@example.com
+
+# File Upload Configuration
+UPLOAD_DIR=uploads
+MAX_FILE_SIZE=10485760  # 10MB in bytes
+ALLOWED_TYPES=image/jpeg,image/png,image/gif,image/webp
+
+# API Configuration
+API_V1_STR=/api/v1
+PROJECT_NAME=LinkLink Image Upload Server
+
+# CORS Configuration
+BACKEND_CORS_ORIGINS=http://localhost:3000,http://localhost:8080
+
+# Logging Configuration
+LOG_LEVEL=INFO
+LOG_FILE=logs/app.log
+
+# Development Configuration
+DEBUG=false 
+
+# Server Configuration
+HOST=0.0.0.0
+PORT=8000
+
+# Docker Configuration
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5432
+POSTGRES_DB=image_upload_db
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=password
