--- test_expired_token.py
@@ -1,107 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script to verify expired token handling
-"""
-
-
-import requests
-
-BASE_URL = "http://localhost:8000"
-API_PREFIX = "/api/v1"
-
-
-def test_expired_token_handling():
-    """Test that expired tokens properly trigger refresh mechanism"""
-    print("🧪 Testing Expired Token Handling")
-    print("=" * 50)
-
-    # 1. Login to get valid tokens
-    print("\n1. Logging in with admin user...")
-    login_data = {"username": "admin", "password": "admin123"}
-
-    response = requests.post(f"{BASE_URL}{API_PREFIX}/login", json=login_data)
-    print(f"Login response: {response.status_code}")
-
-    if response.status_code != 200:
-        print(f"Login error: {response.text}")
-        return False
-
-    login_result = response.json()
-    access_token = login_result.get("access_token")
-    response.cookies
-    print(f"Got access token: {access_token[:20]}...")
-
-    # 2. Test home page with valid token (should show all posts)
-    print("\n2. Testing home page with valid token...")
-    headers = {"Authorization": f"Bearer {access_token}"}
-    response = requests.get(f"{BASE_URL}{API_PREFIX}/posters/", headers=headers)
-    print(f"Home page response: {response.status_code}")
-
-    if response.status_code == 200:
-        posts = response.json()
-        print(f"Number of posts returned: {len(posts)}")
-        if posts:
-            print(f"First post privacy: {posts[0].get('privacy', 'unknown')}")
-
-    # 3. Test with expired token (manually create an expired token)
-    print("\n3. Testing with expired token...")
-    # Create a token that expires in 1 second
-    expired_token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImlzX2FkbWluIjpmYWxzZSwiZXhwIjoxNzM0NzI4MDAwfQ.invalid_signature"
-
-    # Let's test with a completely invalid token first
-    invalid_token = "invalid_token_here"
-    headers = {"Authorization": f"Bearer {invalid_token}"}
-
-    response = requests.get(f"{BASE_URL}{API_PREFIX}/posters/", headers=headers)
-    print(f"Home page with invalid token response: {response.status_code}")
-
-    if response.status_code == 401:
-        print("✅ Correctly returned 401 for invalid token")
-    else:
-        print(f"❌ Expected 401, got {response.status_code}")
-        print(f"Response: {response.text}")
-
-    # Now test with the expired token
-    headers = {"Authorization": f"Bearer {expired_token}"}
-
-    response = requests.get(f"{BASE_URL}{API_PREFIX}/posters/", headers=headers)
-    print(f"Home page with expired token response: {response.status_code}")
-
-    if response.status_code == 401:
-        print("✅ Correctly returned 401 for expired token")
-        return True
-    else:
-        print(f"❌ Expected 401, got {response.status_code}")
-        print(f"Response: {response.text}")
-        return False
-
-
-def test_frontend_refresh_mechanism():
-    """Test the frontend refresh mechanism"""
-    print("\n🧪 Testing Frontend Refresh Mechanism")
-    print("=" * 50)
-
-    # This test would require a browser or frontend testing framework
-    # For now, we'll just verify the backend behavior
-    print("Frontend refresh mechanism test requires browser testing")
-    print("The backend changes should now properly return 401 for expired tokens")
-    print("which will trigger the frontend's fetchWithAuth refresh mechanism")
-
-    return True
-
-
-if __name__ == "__main__":
-    try:
-        success1 = test_expired_token_handling()
-        success2 = test_frontend_refresh_mechanism()
-
-        if success1 and success2:
-            print("\n✅ Expired token handling test completed successfully!")
-            print("\n📝 Summary:")
-            print("- Backend now properly returns 401 for expired tokens")
-            print("- Frontend fetchWithAuth will catch 401 and trigger refresh")
-            print("- Home page will show all posts after successful refresh")
-        else:
-            print("\n❌ Some tests failed!")
-    except Exception as e:
-        print(f"\n❌ Test failed with exception: {e}")
